create configs for dimension props
